backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.27
gocardless-pro==1.46.0
supabase==2.10.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
composio-core==0.6.0
composio-openai==0.6.0